pyproject.toml
[project]
name = "aoc2024"
version = "0.1.0"
description = "My Advent of Code 2024 solutions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "networkx>=3.4.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "watchfiles>=1.0.0",
    "z3-solver>=4.13.3.0",
]

[project.scripts]
aoc2024 = "aoc2024.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]
